+++ spec/javascripts/collectionView.spec.js
@@ -197,7 +197,6 @@ describe("collection view", function(){
       emptyView: EmptyView
     });
 
-
     describe("when rendering a collection view with an empty collection", function(){
 
       var collectionView;
@@ -340,6 +339,46 @@ describe("collection view", function(){
       expect(collectionView.appendHtml.calls[0].args[2]).toBe(0);
     });
 
+  });
+
+  describe("when providing a custom render that adds children, without a collection object to use, and removing a child", function(){
+    var collectionView;
+    var childView;
+
+    var model = new Backbone.Model({foo: "bar"});
+
+    var EmptyView = Backbone.Marionette.ItemView.extend({
+      render: function(){}
+    });
+
+    var CollectionView = Backbone.Marionette.CollectionView.extend({
+      itemView: ItemView,
+      emptyView: EmptyView,
+
+      render: function(){
+        var ItemView = this.getItemView();
+        this.addItemView(model, ItemView, 0);
+      }
+    });
+
+    beforeEach(function(){
+      collectionView = new CollectionView({});
+      collectionView.render();
+
+      childView = collectionView.children[model.cid];
+      spyOn(childView, "close").andCallThrough();
+      spyOn(EmptyView.prototype, "render");
+
+      collectionView.removeItemView(model);
+    });
+
+    it("should close the model's view", function(){
+      expect(childView.close).toHaveBeenCalled();
+    });
+
+    it("should show the empty view", function(){
+      expect(EmptyView.prototype.render.callCount).toBe(1);
+    });
   });
 
   describe("when a model is removed from the collection", function(){