+++ changelog.md
@@ -7,6 +7,9 @@
   * Defer module definition until module is started
   * App.start sends options to module initializers
 
+* CollectionView
+  * Correctly handles non-existent collection and removing child item views that were added manually
+
 ### v0.9.3 [view commit logs](https://github.com/derickbailey/backbone.marionette/compare/v0.9.2...v0.9.3)
 
 * CompositeView